pyproject.toml
[project]
name = "messaging-service"
version = "0.1.0"
description = "WhatsApp messaging platform service"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1,<7.0",
    "djangorestframework>=3.15",
    "django-environ>=0.11",
    "gunicorn>=22.0",
    "psycopg[binary]>=3.1",
    "whitenoise>=6.5",
    "httpx>=0.27",
    "redis>=5.0",
    "celery[redis]>=5.4",
    "platform-contracts @ git+https://github.com/dziadekp/platform-contracts.git@v0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I"]
